Move validate step and fix docstring
